+++ src/app/experience-list/experience-list.component.html
@@ -21,6 +21,6 @@
   <form>
     <h2>Add Experience</h2>
     <input type="text" placeholder="Position Title" #newExperience>
-    <button (click)="saveExperience(newExperience.value)">Add</button>
+    <button type="button" (click)="saveExperience(newExperience.value)">Add</button>
   </form>
 </section>
